feat(workflows): add Docker build and push workflow
This commit adds a new GitHub Actions workflow named "Docker Build
and Push".

The workflow is triggered by a workflow call and requires
several inputs and secrets.

It includes steps to check and create an artifact repository, set the
Docker image, authenticate with Docker, build, tag, and push the Docker
image, and manage build cache.