feat(workflows): add Google Cloud auth workflow
This commit introduces a new GitHub Actions workflow named "Google
Cloud Authentication".

The workflow is triggered by a workflow call and requires several
secrets for Google Cloud authentication.

It includes a job that runs on the latest Ubuntu environment and uses
the google-github-actions/auth action to obtain an access token.